reactor module refactored to package.
todo: add new features to Reactor.
