+++ CGRtools/reactor/base.py
@@ -18,15 +18,9 @@
 #  along with this program; if not, see <https://www.gnu.org/licenses/>.
 #
 from collections import defaultdict
-from functools import reduce
-from itertools import count, permutations
-from logging import info
-from operator import or_
-from typing import Union, Iterable
-from ._functions import lazy_product
-from .containers import QueryContainer, QueryCGRContainer, MoleculeContainer, CGRContainer, ReactionContainer
-from .containers.bonds import Bond
-from .periodictable import Element, DynamicElement
+from ..containers import QueryContainer
+from ..containers.bonds import Bond
+from ..periodictable import Element, DynamicElement
 
 
 class BaseReactor:
@@ -237,149 +231,4 @@ class BaseReactor:
         self.__bonds = state['bonds']
 
 
-class CGRReactor(BaseReactor):
-    """
-    CGR based editor for CGRs and molecules.
-    generates transformation from input CGR/molecule
-    using template (CGRtools ReactionContainer).
-    Template should contain one reactant and one product:
-
-    CGRReactor calling transforms reactants to products and
-    returns generator of all possible products.
-    """
-
-    def __init__(self, template: ReactionContainer, delete_atoms: bool = True):
-        """
-        :param template: CGRtools ReactionContainer
-        :param delete_atoms: if True atoms exists in reactant but
-                            not exists in product will be removed
-        """
-        reactants, products = template.reactants, template.products
-        if not reactants or not products:
-            raise ValueError('empty template')
-        if isinstance(reactants[0], CGRContainer):
-            reactants = reduce(or_, reactants, QueryCGRContainer())
-            products = reduce(or_, products, QueryCGRContainer())
-        elif isinstance(reactants[0], QueryCGRContainer):
-            reactants = reduce(or_, reactants)
-            products = reduce(or_, products)
-        elif isinstance(reactants[0], MoleculeContainer):
-            reactants = reduce(or_, reactants, QueryContainer())
-            products = reduce(or_, products, QueryContainer())
-        else:
-            reactants = reduce(or_, reactants)
-            products = reduce(or_, products)
-
-        self.__pattern = reactants
-        self.__meta = template.meta.copy()
-        super().__init__(reactants, products, delete_atoms)
-
-    def __call__(self, structure: Union[MoleculeContainer, CGRContainer], automorphism_filter: bool = True):
-        if not isinstance(structure, (MoleculeContainer, CGRContainer)):
-            raise TypeError('only Molecules and CGRs possible')
-
-        for mapping in self.__pattern.get_mapping(structure, automorphism_filter=automorphism_filter):
-            new = self._patcher(structure, mapping)
-            new.meta.update(self.__meta)
-            yield new
-
-    def __getstate__(self):
-        return {'pattern': self.__pattern, 'meta': self.__meta, **super().__getstate__()}
-
-    def __setstate__(self, state):
-        self.__pattern = state['pattern']
-        self.__meta = state['meta']
-        super().__setstate__(state)
-
-
-class Reactor(BaseReactor):
-    """
-    CGR based reactor for molecules/queries.
-    generates reaction from input queries/molecules using
-    transformation template (CGRtools ReactionContainer).
-
-    Reactor calling transforms reactants to products and
-    returns generator of reaction transformations with all
-    possible products
-    """
-
-    def __init__(self, template, delete_atoms=True):
-        """
-        :param template: CGRtools ReactionContainer
-        :param delete_atoms: if True atoms exists in reactants but
-                            not exists in products will be removed
-        """
-        reactants, products = template.reactants, template.products
-        if not reactants or not products:
-            raise ValueError('empty template')
-
-        self.__split = len(products)
-
-        if isinstance(reactants[0], MoleculeContainer):
-            self.__patterns = reactants = tuple(QueryContainer() | x for x in reactants)
-            products = reduce(or_, products, QueryContainer())
-        elif isinstance(reactants[0], QueryContainer):
-            self.__patterns = reactants
-            products = reduce(or_, products)
-        else:
-            raise TypeError('only Molecules and Queries possible')
-
-        reactants = reduce(or_, reactants)
-        self.__meta = template.meta.copy()
-        super().__init__(reactants, products, delete_atoms)
-
-    def __call__(self, structures: Iterable[MoleculeContainer], automorphism_filter: bool = True):
-        if any(not isinstance(structure, MoleculeContainer) for structure in structures):
-            raise TypeError('only list of Molecules possible')
-
-        structures = self.__remap(structures)
-        s_nums = set(range(len(structures)))
-        for chosen in permutations(s_nums, len(self.__patterns)):
-            ignored = [structures[x] for x in s_nums.difference(chosen)]
-            ignored_numbers = {x for x in ignored for x in x}
-            chosen = [structures[x] for x in chosen]
-            united_chosen = reduce(or_, chosen)
-            for match in lazy_product(*(x.get_mapping(y, automorphism_filter=automorphism_filter) for x, y in
-                                        zip(self.__patterns, chosen))):
-                mapping = match[0]
-                for m in match[1:]:
-                    mapping.update(m)
-                new = self._patcher(united_chosen, mapping)
-                collision = set(new).intersection(ignored_numbers)
-                if collision:
-                    new.remap(dict(zip(collision, count(max(max(ignored_numbers), max(new.atoms_numbers)) + 1))))
-                if self.__split > 1:
-                    new = new.split()
-                    if len(new) != self.__split:
-                        info(f'expected {self.__split} molecules in reaction products, but {len(new)} formed.\n'
-                             'input molecules has disconnected components')
-                else:
-                    new = [new]
-                yield ReactionContainer(structures, new + ignored, meta=self.__meta)
-
-    @staticmethod
-    def __remap(structures):
-        checked = []
-        checked_atoms = set()
-        for structure in structures:
-            intersection = set(structure).intersection(checked_atoms)
-            if intersection:
-                mapping = dict(zip(intersection, count(max(max(checked_atoms), max(structure.atoms_numbers)) + 1)))
-                structure = structure.remap(mapping, copy=True)
-                info('some atoms in input structures had the same numbers.\n'
-                     f'atoms {list(mapping)} were remapped to {list(mapping.values())}')
-            checked_atoms.update(structure)
-            checked.append(structure)
-        return checked
-
-    def __getstate__(self):
-        return {'patterns': self.__patterns, 'meta': self.__meta, 'split': self.__split, **super().__getstate__()}
-
-    def __setstate__(self, state):
-        self.__patterns = state['patterns']
-        self.__meta = state['meta']
-        self.__split = state['split']
-        super().__setstate__(state)
-
-
-__all__ = ['CGRReactor', 'Reactor']
+__all__ = ['BaseReactor']
